bean validation

+++ src/main/java/com/tecmulti/bookmanagement/service/BookService.java
@@ -1,14 +1,18 @@
 package com.tecmulti.bookmanagement.service;
 
+import org.hibernate.ResourceClosedException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.tecmulti.bookmanagement.dto.BookDto;
 import com.tecmulti.bookmanagement.entities.Book;
 import com.tecmulti.bookmanagement.repositories.BookRepository;
+import com.tecmulti.bookmanagement.service.exeption.DataBaseExeption;
 import com.tecmulti.bookmanagement.service.exeption.ResourceNotFountExeption;
 
 @Service
@@ -45,9 +49,17 @@ public class BookService {
 		return new BookDto(books);
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(! repository.existsById(id)){
+			 throw new ResourceNotFountExeption("Recurso não encontrado");
+		}
+		try {
+			repository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DataBaseExeption("falha de integridade");
+		}
 	}
 	
 	private static Book dtoToEntity(Book book, BookDto dto) {//tirar repetição de update e insert
